+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added config for starter items when entering the Nether for the first time
+- Added grace period with buffs when entering the Nether for the first time
 
 ### Fixed